+++ Aspire/MC.Aspire.AppHost/AppHost.cs
@@ -47,7 +47,7 @@ static void AddDeveloperGraph(IDistributedApplicationBuilder builder)
         .WithReference(catalogMongoDatabase, catalogMongoDatabaseName);
 
     // React Web App
-    var reactwebapp = builder.AddViteApp("reactWebApp", "./../../Clients/MC.Market.WebApp", "dev");
+    var reactwebapp = builder.AddViteApp("reactWebApp", "./../../Clients/mc.market.webapp", "dev");
 
     // Gateway
     var webStoreGateway = builder.AddProject<Projects.MC_Gateway>("marketCustomsGateway")