+++ Clients/MC.Market.WebApp old/src/main.tsx
@@ -1,0 +1,20 @@
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
+import App from "./App.tsx";
+import { Auth0Provider } from "@auth0/auth0-react";
+import { auth0Config } from "./auth/auth0.ts";
+
+createRoot(document.getElementById("root")!).render(
+  <StrictMode>
+    <Auth0Provider
+      domain={auth0Config.domain ?? ""}
+      clientId={auth0Config.clientId ?? ""}
+      authorizationParams={auth0Config.authorizationParams}
+      cacheLocation="localstorage"
+      useRefreshTokens
+    >
+      <App />
+    </Auth0Provider>
+  </StrictMode>,
+);