+++ Clients/MC.Market.WebApp/src/main.tsx
@@ -1,20 +1,10 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import "./index.css";
-import App from "./App.tsx";
-import { Auth0Provider } from "@auth0/auth0-react";
-import { auth0Config } from "./auth/auth0.ts";
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import './index.css'
+import App from './App.tsx'
 
-createRoot(document.getElementById("root")!).render(
+createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <Auth0Provider
-      domain={auth0Config.domain ?? ""}
-      clientId={auth0Config.clientId ?? ""}
-      authorizationParams={auth0Config.authorizationParams}
-      cacheLocation="localstorage"
-      useRefreshTokens
-    >
-      <App />
-    </Auth0Provider>
+    <App />
   </StrictMode>,
-);
+)