+++ Clients/MC.Market.WebApp old/CHANGELOG.md
@@ -4,9 +4,9 @@ The following tools were used to generate this project:
 - create-vite
 
 The following steps were used to generate this project:
-- Create react project with create-vite: `npm init --yes vite@latest mc.market.webapp -- --template=react-ts  --no-rolldown --no-immediate`.
+- Create react project with create-vite: `npm init --yes vite@latest mc.market.reactwebapp -- --template=react-ts  --no-rolldown --no-immediate`.
 - Updating `vite.config.ts` with port.
-- Create project file (`mc.market.webapp.esproj`).
+- Create project file (`mc.market.reactwebapp.esproj`).
 - Create `launch.json` to enable debugging.
 - Add project to solution.
 - Write this file.